services/user_service/requirements.txt
django==5.1.0 
djangorestframework==3.15.2 
djangorestframework-simplejwt==5.3.1 
drf-social-oauth2==1.1.3
django-allauth==0.57.0
psycopg2-binary==2.9.9 
gunicorn==23.0.0 
django-environ==0.11.2
pika==1.3.2
python-consul==1.1.0
Pillow==10.4.0
django-cors-headers==4.4.0
google-auth==2.30.0
requests-oauthlib==1.3.1
aiohttp==3.9.1